fix: Pull request feedback
Moved account.assetBulkOptIn -> asset.bulkOptIn
Moved account.assetBulkOptOut -> asset.bulkOptOut
Moved account.getAssetInformation -> asset.getAccountInformation
Fixed other miscellaneous issues
