--- docs/code/interfaces/types_account_manager.BulkAssetOptInOutResult.md
@@ -1,38 +1,0 @@
-[@algorandfoundation/algokit-utils](../README.md) / [types/account-manager](../modules/types_account_manager.md) / BulkAssetOptInOutResult
-
-# Interface: BulkAssetOptInOutResult
-
-[types/account-manager](../modules/types_account_manager.md).BulkAssetOptInOutResult
-
-Individual result from performing a bulk opt-in or bulk opt-out for an account against a series of assets.
-
-## Table of contents
-
-### Properties
-
-- [assetId](types_account_manager.BulkAssetOptInOutResult.md#assetid)
-- [transactionId](types_account_manager.BulkAssetOptInOutResult.md#transactionid)
-
-## Properties
-
-### assetId
-
-• **assetId**: `bigint`
-
-The ID of the asset opted into / out of
-
-#### Defined in
-
-[src/types/account-manager.ts:27](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/account-manager.ts#L27)
-
-___
-
-### transactionId
-
-• **transactionId**: `string`
-
-The transaction ID of the resulting opt in / out
-
-#### Defined in
-
-[src/types/account-manager.ts:29](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/account-manager.ts#L29)
