+++ docs/code/interfaces/types_asset_manager.BulkAssetOptInOutResult.md
@@ -1,0 +1,38 @@
+[@algorandfoundation/algokit-utils](../README.md) / [types/asset-manager](../modules/types_asset_manager.md) / BulkAssetOptInOutResult
+
+# Interface: BulkAssetOptInOutResult
+
+[types/asset-manager](../modules/types_asset_manager.md).BulkAssetOptInOutResult
+
+Individual result from performing a bulk opt-in or bulk opt-out for an account against a series of assets.
+
+## Table of contents
+
+### Properties
+
+- [assetId](types_asset_manager.BulkAssetOptInOutResult.md#assetid)
+- [transactionId](types_asset_manager.BulkAssetOptInOutResult.md#transactionid)
+
+## Properties
+
+### assetId
+
+• **assetId**: `bigint`
+
+The ID of the asset opted into / out of
+
+#### Defined in
+
+[src/types/asset-manager.ts:13](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/asset-manager.ts#L13)
+
+___
+
+### transactionId
+
+• **transactionId**: `string`
+
+The transaction ID of the resulting opt in / out
+
+#### Defined in
+
+[src/types/asset-manager.ts:15](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/asset-manager.ts#L15)
